[api] Add backoff to database connect call.
Add an exponential backoff to the get_db_conn call to make it retry: The
Docker Compose "depends_on" is not enough to ensure that the database is
available before the api application tries to use it.

NOTE: The backoff "max_time" arg is set to 60 seconds, which is higher
than the default gunicorn worker timeout of 30 seconds. This makes
gunicorn restart the worker every 30 seconds while waiting for the
database connection.

If "max_time" is set lower than gunicorn's timeout, the worker will
raise an exception that causes the entire application to exit.